Removed script wrapper.
The code is really intended to be used as a library so providing entry
points that turn each function into a script that can be run on the
command line is overkill. Plus the scripts are very simply so it
pollutes the system path with a lot of extra executables.