Removed unit tests.
The unit tests used snippets of HTML extracted from a real checklist.
That worked great except when the styling of a page changed radically.
Then all the tests needed to be updated with the changed HTML. That
sort of worked when the pages were relatively small and probably
hand-written. With the latest version (2019) the volume of HTML is
large and the semantic structure has been eliminated and instead
functional css is used. That makes it a lot harded to extract small
snippets of HTML since the data for a given field is now in the last
<li> item in the fourth <div> of the third <section>. The obvious
solution is to simply maintain a complete page and parse it but that
does not solve the problem of knowing that the HTML on the site
changed and now the code does not work. So the solution is to create
a set of test checklists on eBird itself and then run the tests
against those pages. If the checklists are minimal and for private
locations then it will only pollute the database by a very tiny amount
given the solume of records added. That also makes it easier to run
checks periodically to see if the code is up to date. The checklists
and corresponding tests will be done later.