Updated get_checklist to work with latest HTML.
The get_checklist script now works with the latest version (Dec 2019) of
the view checklist page. The changes are just enough to make it work and
do not support all the features of the previous release - the most obvious
one is "observers". The code has only been tested with the incidental and
travelling protocols but should work for others.

This is just an interim commit to get the project back to a working state.
It's not sufficient for a release and more updates will be added soon.